+++ Mac/Sources/fe-aqua_common.h
@@ -46,4 +46,4 @@ void change_channel_flag (struct session *sess, char flag, int enabled);
 void set_l_flag (struct session *sess, int enabled, int value);
 void set_k_flag (struct session *sess, int enabled, char *value);
 
-NSString * formatNumber (int n);
+NSString *formatNumber(NSInteger n);
